Factor out frontend/legacy_forms.py.
