+++ frontend/views.py
@@ -9,7 +9,6 @@ from django.urls import reverse
 from django.conf import settings
 from django.contrib.sites.models import Site
 from users.models import JustfixUser
-from project.util import django_graphql_forms
 from project.justfix_environment import BASE_DIR
 from project.util.lambda_service import LambdaService
 from project.util.site_util import (
@@ -18,16 +17,15 @@ from project.util.site_util import (
     get_site_origin,
 )
 from project.graphql_static_request import GraphQLStaticRequest
-from project import common_data
+
 from .graphql import execute_query, get_initial_session
 from .lambda_response import GraphQLQueryPrefetchInfo, LambdaResponse
+from .legacy_forms import LegacyFormSubmissionError, get_legacy_form_submission
 
 # This is changed by test suites to ensure that
 # everything works okay when the server-side renderer fails
 # (relatively) gracefully.
 TEST_INTERNAL_SERVER_ERROR = False
-
-FORMS_COMMON_DATA = common_data.load_json("forms.json")
 
 NS_PER_MS = 1e+6
 
@@ -115,53 +113,6 @@ def run_react_lambda_with_prefetching(initial_props, request) -> LambdaResponse:
         )
 
     return lambda_response
-
-
-class LegacyFormSubmissionError(Exception):
-    pass
-
-
-def fix_newlines(d: Dict[str, str]) -> Dict[str, str]:
-    result = dict()
-    result.update(d)
-    for key in d:
-        result[key] = result[key].replace('\r\n', '\n')
-    return result
-
-
-def get_legacy_form_submission_result(request, graphql, input):
-    if request.POST.get(FORMS_COMMON_DATA["LEGACY_FORMSET_ADD_BUTTON_NAME"]):
-        return None
-    return execute_query(request, graphql, variables={'input': input})['output']
-
-
-def get_legacy_form_submission(request) -> Dict[str, Any]:
-    graphql = request.POST.get('graphql')
-
-    if not graphql:
-        raise LegacyFormSubmissionError('No GraphQL query found')
-
-    input_type = django_graphql_forms.get_input_type_from_query(graphql)
-
-    if not input_type:
-        raise LegacyFormSubmissionError('Invalid GraphQL query')
-
-    form_class = django_graphql_forms.get_form_class_for_input_type(input_type)
-
-    if not form_class:
-        raise LegacyFormSubmissionError('Invalid GraphQL input type')
-
-    formset_classes = django_graphql_forms.get_formset_classes_for_input_type(input_type)
-    exclude_fields = django_graphql_forms.get_exclude_fields_for_input_type(input_type)
-
-    input = django_graphql_forms.convert_post_data_to_input(
-        form_class, request.POST, formset_classes, exclude_fields)
-
-    return {
-        'input': input,
-        'result': get_legacy_form_submission_result(request, graphql, input),
-        'POST': fix_newlines(request.POST.dict())
-    }
 
 
 def get_webpack_public_path_url() -> str:
